Move remote registry operations into dedicated module
